add upload button to main page
